refactor(coderd/x/chatd): extract hook integration into chathooks subpackage
Move the lifecycle hook trigger seam, transcript effects, tool gate,
and error classification from chatd into coderd/x/chatd/chathooks,
exporting the moved symbols. Server-bound glue stays in chatd
(hook_server.go): the chat-parking dispatch error handlers, the
step-commit row insertion wrappers, and the dynamic post-tool-use state
loader, which depends on chatd validation types. Trigger-focused
internal tests move with the package; fixture-bound tests stay.
