+++ coderd/x/chatd/chathooks/tooluse.go
@@ -1,11 +1,10 @@
-package chatd
+package chathooks
 
 import (
 	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 
 	"charm.land/fantasy"
 	"github.com/google/uuid"
@@ -35,44 +34,44 @@ func rejectDuplicateToolUseIDs(toolCalls []fantasy.ToolCallContent) error {
 	return nil
 }
 
-// preToolUseExecutionResult preserves hook results in tool-call order for
+// PreToolUseExecutionResult preserves hook results in tool-call order for
 // transcript injection.
-type preToolUseExecutionResult struct {
+type PreToolUseExecutionResult struct {
 	Allowed   []fantasy.ToolCallContent
 	Denied    []fantasy.ToolResultContent
-	Results   []*hookResult
+	Results   []*Result
 	Overrides map[string]json.RawMessage
 }
 
-func (t *hookTrigger) preflightPendingToolCalls(
+func (t *Trigger) PreflightPendingToolCalls(
 	ctx context.Context,
-	chat hookChat,
+	chat Chat,
 	toolCalls []fantasy.ToolCallContent,
-) (preToolUseExecutionResult, error) {
-	if !t.enabled() {
-		return preToolUseExecutionResult{Allowed: toolCalls}, nil
+) (PreToolUseExecutionResult, error) {
+	if !t.Enabled() {
+		return PreToolUseExecutionResult{Allowed: toolCalls}, nil
 	}
-	result := preToolUseExecutionResult{
+	result := PreToolUseExecutionResult{
 		Allowed: make([]fantasy.ToolCallContent, 0, len(toolCalls)),
 	}
 	if err := rejectDuplicateToolUseIDs(toolCalls); err != nil {
-		return preToolUseExecutionResult{}, err
+		return PreToolUseExecutionResult{}, err
 	}
 
 	for _, toolCall := range toolCalls {
-		callResult, err := t.trigger(ctx, chat, hookMessage{
+		callResult, err := t.Trigger(ctx, chat, Message{
 			ToolUseID: toolCall.ToolCallID,
 			ToolName:  toolCall.ToolName,
 			ToolInput: json.RawMessage(toolCall.Input),
 		}, hooks.EventPreToolUse)
 		if err != nil {
-			denied, ok := errors.AsType[*hookDeniedError](err)
+			denied, ok := errors.AsType[*deniedError](err)
 			if !ok {
-				return preToolUseExecutionResult{}, err
+				return PreToolUseExecutionResult{}, err
 			}
 			// The denial's model context folds into the synthetic tool
 			// result; only the user notice needs a transcript row.
-			result.Results = append(result.Results, &hookResult{UserMessage: denied.UserMessage})
+			result.Results = append(result.Results, &Result{UserMessage: denied.UserMessage})
 			result.Denied = append(result.Denied, deniedToolResult(toolCall, denied.Reason, denied.ModelContext))
 			continue
 		}
@@ -89,8 +88,8 @@ func (t *hookTrigger) preflightPendingToolCalls(
 	return result, nil
 }
 
-func postToolUseMessage(toolResult fantasy.ToolResultContent) (hookMessage, error) {
-	msg := hookMessage{
+func postToolUseMessage(toolResult fantasy.ToolResultContent) (Message, error) {
+	msg := Message{
 		ToolUseID: toolResult.ToolCallID,
 		ToolName:  toolResult.ToolName,
 	}
@@ -106,22 +105,22 @@ func postToolUseMessage(toolResult fantasy.ToolResultContent) (hookMessage, erro
 	default:
 		encoded, err := json.Marshal(toolResult.Result)
 		if err != nil {
-			return hookMessage{}, xerrors.Errorf("marshal post_tool_use response: %w", err)
+			return Message{}, xerrors.Errorf("marshal post_tool_use response: %w", err)
 		}
 		msg.ToolResponse = encoded
 	}
 	return msg, nil
 }
 
-func (t *hookTrigger) postToolUseResults(
+func (t *Trigger) PostToolUseResults(
 	ctx context.Context,
-	chat hookChat,
+	chat Chat,
 	content []fantasy.Content,
-) ([]*hookResult, error) {
-	if !t.enabled() {
+) ([]*Result, error) {
+	if !t.Enabled() {
 		return nil, nil
 	}
-	results := make([]*hookResult, 0, len(content))
+	results := make([]*Result, 0, len(content))
 	var firstErr error
 	for _, block := range content {
 		toolResult, ok := asToolResultContent(block)
@@ -135,7 +134,7 @@ func (t *hookTrigger) postToolUseResults(
 			}
 			continue
 		}
-		result, err := t.trigger(ctx, chat, msg, hooks.EventPostToolUse)
+		result, err := t.Trigger(ctx, chat, msg, hooks.EventPostToolUse)
 		if err != nil {
 			if firstErr == nil {
 				firstErr = err
@@ -147,7 +146,7 @@ func (t *hookTrigger) postToolUseResults(
 	return results, firstErr
 }
 
-func replacePersistedToolCallInputs(
+func ReplacePersistedToolCallInputs(
 	ctx context.Context,
 	tx *chatstate.Tx,
 	chatID uuid.UUID,
@@ -187,97 +186,8 @@ func replacePersistedToolCallInputs(
 	return nil
 }
 
-type dynamicPostToolUseState struct {
-	chat          database.Chat
-	modelConfigID uuid.UUID
-	toolNames     map[string]string
-}
-
-func loadDynamicPostToolUseState(
-	ctx context.Context,
-	machine *chatstate.ChatMachine,
-	opts SubmitToolResultsOptions,
-) (dynamicPostToolUseState, error) {
-	var state dynamicPostToolUseState
-	err := machine.ReadLock(ctx, func(store database.Store) error {
-		chat, err := store.GetChatByID(ctx, opts.ChatID)
-		if err != nil {
-			return xerrors.Errorf("load chat: %w", err)
-		}
-		if chat.Archived {
-			return ErrChatArchived
-		}
-		if chat.Status != database.ChatStatusRequiresAction {
-			return &ToolResultStatusConflictError{ActualStatus: chat.Status}
-		}
-		messages, err := store.GetChatMessagesByChatID(ctx, database.GetChatMessagesByChatIDParams{
-			ChatID:  opts.ChatID,
-			AfterID: 0,
-		})
-		if err != nil {
-			return xerrors.Errorf("load chat messages: %w", err)
-		}
-		_, pending, err := unresolvedToolCallsFromHistory(messages, dynamicToolNamesFromChat(chat))
-		if err != nil {
-			return xerrors.Errorf("load pending dynamic tool calls: %w", err)
-		}
-		toolNames := make(map[string]string, len(pending))
-		for _, call := range pending {
-			toolNames[call.ToolCallID] = call.ToolName
-		}
-		if err := validateSubmittedToolResults(opts.Results, toolNames); err != nil {
-			return err
-		}
-		modelConfigID := opts.ModelConfigID
-		if modelConfigID == uuid.Nil {
-			modelConfigID = chat.LastModelConfigID
-		}
-		state = dynamicPostToolUseState{
-			chat:          chat,
-			modelConfigID: modelConfigID,
-			toolNames:     toolNames,
-		}
-		return nil
-	})
-	return state, err
-}
-
-func validateSubmittedToolResults(results []codersdk.ToolResult, toolNames map[string]string) error {
-	submitted := make(map[string]struct{}, len(results))
-	for _, result := range results {
-		if _, ok := submitted[result.ToolCallID]; ok {
-			return &ToolResultValidationError{
-				Message: "Duplicate tool_call_id in results.",
-				Detail:  fmt.Sprintf("Duplicate tool call ID %q.", result.ToolCallID),
-			}
-		}
-		if !json.Valid(result.Output) {
-			return &ToolResultValidationError{
-				Message: "Tool result output must be valid JSON.",
-				Detail:  fmt.Sprintf("Output for tool call %q is not valid JSON.", result.ToolCallID),
-			}
-		}
-		if _, ok := toolNames[result.ToolCallID]; !ok {
-			return &ToolResultValidationError{
-				Message: "Unexpected tool result.",
-				Detail:  fmt.Sprintf("No pending tool call with ID %q.", result.ToolCallID),
-			}
-		}
-		submitted[result.ToolCallID] = struct{}{}
-	}
-	for toolCallID := range toolNames {
-		if _, ok := submitted[toolCallID]; !ok {
-			return &ToolResultValidationError{
-				Message: "Missing tool result.",
-				Detail:  fmt.Sprintf("Missing result for tool call %q.", toolCallID),
-			}
-		}
-	}
-	return nil
-}
-
-func dynamicPostToolUseMessage(result codersdk.ToolResult, toolName string) hookMessage {
-	msg := hookMessage{
+func DynamicPostToolUseMessage(result codersdk.ToolResult, toolName string) Message {
+	msg := Message{
 		ToolUseID: result.ToolCallID,
 		ToolName:  toolName,
 	}
